Move fetching logic to useAdvice hook
